adapted for long

+++ src/frontend/org/qcri/PartitioningPlanner/placement/FirstFitPlacement.java
@@ -11,14 +11,14 @@ import org.qcri.PartitioningPlanner.placement.Plan;
 
 public class FirstFitPlacement extends Placement {
 	
-	Integer coldPartitionWidth = 1000; // redistribute cold tuples in chunks of 1000
+	Long coldPartitionWidth = 1000L; // redistribute cold tuples in chunks of 1000
 	
 	public FirstFitPlacement(){
 		
 	}
 	
-	static Integer getMostUnderloadedPartitionId(Map<Integer, Integer> partitionTotals) {
-		Integer minTotal = java.lang.Integer.MAX_VALUE; 
+	static Integer getMostUnderloadedPartitionId(Map<Integer, Long> partitionTotals) {
+		Long minTotal = java.lang.Long.MAX_VALUE; 
 		Integer minPartition = -1;
 
 		for(Integer i : partitionTotals.keySet()) {
@@ -34,13 +34,13 @@ public class FirstFitPlacement extends Placement {
 	
 	// hotTuples: tupleId --> access count
 	// siteLoads: partitionId --> total access count
-	public Plan computePlan(ArrayList<Map<Integer, Integer>> hotTuplesList, Map<Integer, Integer> partitionTotals, Plan aPlan){
+	public Plan computePlan(ArrayList<Map<Long, Long>> hotTuplesList, Map<Integer, Long> partitionTotals, Plan aPlan){
 		
 
 		Integer srcPartition = 0, dstPartition = -1;
-		Integer totalAccesses = 0;
-		Integer targetCapacity;
-		Map<Integer, Integer> oldLoad = new HashMap<Integer, Integer> ();
+		Long totalAccesses = 0L;
+		Long targetCapacity;
+		Map<Integer, Long> oldLoad = new HashMap<Integer, Long> ();
 		Plan newPlan = new Plan();
 		
 
@@ -48,7 +48,7 @@ public class FirstFitPlacement extends Placement {
 			totalAccesses = totalAccesses + partitionTotals.get(i);			
 			oldLoad.put(i,  partitionTotals.get(i));
 			// zero out the load for a plan
-			partitionTotals.put(i, 0);
+			partitionTotals.put(i, 0L);
 		}
 
 		
@@ -60,9 +60,9 @@ public class FirstFitPlacement extends Placement {
 		}
 		
 		// pack the hot tuples first
-		for(Map<Integer, Integer> hotTuples : hotTuplesList) {
+		for(Map<Long, Long> hotTuples : hotTuplesList) {
 			
-			for(Integer i : hotTuples.keySet()) {
+			for(Long i : hotTuples.keySet()) {
 				Boolean placed = false;
 				for(Integer j : partitionTotals.keySet()) {
 					if(partitionTotals.get(j) + hotTuples.get(i) <= targetCapacity) {
